requirements.txt
# Aura AI - Unified Requirements
# Core
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0

# AI / ML
numpy==1.26.3
opencv-python==4.9.0.80
torch==2.1.2
transformers==4.37.2
huggingface_hub==0.20.3
timm==0.9.12
pillow==10.2.0

# Frontend
streamlit==1.30.0
audio-recorder-streamlit==0.0.8
altair==5.2.0

# Utilities
pydantic==2.5.3
aiofiles==23.2.1
